requirements.txt
fastapi
uvicorn
opencv-python
numpy
pillow
mediapipe
scikit-image
python-multipart
streamlit
tensorflow==2.12  # if you DO NOT use CNN, remove this
